+++ src/structures/node.jl
@@ -225,12 +225,19 @@ struct PinchData{
 end
 
 """
+    AbstractTES <: Storage{T}
+
+Abstract supertype for all thermal energy storage nodes.
+"""
+abstract type AbstractTES{T} <: Storage{T} end
+
+"""
     ThermalEnergyStorage{T} <: Storage{T}
 
 A `ThermalEnergyStorage` that functions mostly like a [`RefStorage`](@extref EnergyModelsBase.RefStorage)
-with the additional option to include thermal energy losses. Heat losses are quantified
-through a heat loss factor that describes the amount of thermal energy that is lost in
-relation to the storage level from the previous timeperiod.
+with the additional option to include a discharge rate and thermal energy losses. Heat losses
+are quantified through a heat loss factor that describes the amount of thermal energy that is
+lost in relation to the storage level from the previous timeperiod.
 
 The main difference to [`RefStorage`](@extref EnergyModelsBase.RefStorage) is that these
 heat losses do not occur while charging or discharging, *i.e.*, they are proportional to the
@@ -248,7 +255,10 @@ storage level.
   include variable OPEX, fixed OPEX, and/or a capacity.
 - **`level::AbstractStorageParameters`** are the level parameters of the `ThermalEnergyStorage`.
   Depending on the chosen type, the charge parameters can include variable OPEX and/or fixed OPEX.
-- **`stor_res::Resource`** is the stored [`Resource`](@extref EnergyModelsBase.Resource).
+- **`discharge::AbstractStorageParameters`** are the discharging parameters of the
+  `ThermalEnergyStorage` node. Depending on the chosen type, the discharge parameters can
+  include variable OPEX, fixed OPEX, and/or a capacity.
+  - **`stor_res::Resource`** is the stored [`Resource`](@extref EnergyModelsBase.Resource).
 - **`heat_loss_factor::Float64`** are the relative heat losses in percent.
 - **`input::Dict{<:Resource,<:Real}`** are the input [`Resource`](@extref EnergyModelsBase.Resource)s
   with conversion value `Real`.
@@ -259,10 +269,11 @@ storage level.
 - **`data::Vector{<:ExtensionData}`** is the additional data (*e.g.*, for investments). The
   field `data` is conditional through usage of a constructor.
 """
-struct ThermalEnergyStorage{T} <: Storage{T}
+struct ThermalEnergyStorage{T} <: AbstractTES{T}
     id::Any
     charge::EMB.AbstractStorageParameters
     level::EMB.UnionCapacity
+    discharge::EMB.AbstractStorageParameters
     stor_res::Resource
     heat_loss_factor::Float64
     input::Dict{<:Resource,<:Real}
@@ -274,6 +285,7 @@ function ThermalEnergyStorage{T}(
     id,
     charge::EMB.AbstractStorageParameters,
     level::EMB.UnionCapacity,
+    discharge::EMB.AbstractStorageParameters,
     stor_res::Resource,
     heat_loss_factor::Float64,
     input::Dict{<:Resource,<:Real},
@@ -283,6 +295,7 @@ function ThermalEnergyStorage{T}(
         id,
         charge,
         level,
+        discharge,
         stor_res,
         heat_loss_factor,
         input,
@@ -295,6 +308,7 @@ function ThermalEnergyStorage(
     id::Any,
     charge::EMB.AbstractStorageParameters,
     level::EMB.UnionCapacity,
+    discharge::EMB.AbstractStorageParameters,
     stor_res::Resource,
     heat_loss_factor::Float64,
     input::Dict{<:Resource,<:Real},
@@ -305,8 +319,97 @@ function ThermalEnergyStorage(
         id,
         charge,
         level,
+        discharge,
         stor_res,
         heat_loss_factor,
+        input,
+        output,
+        data,
+    )
+end
+
+"""
+    BoundRateTES{T} <: AbstractTES{T}
+
+A `BoundRateTES` that has the option to include thermal energy losses. In contrast to
+[`ThermalEnergyStorage`](@ref), the maximum charging and discharging rates are defined as a
+ratio between the (dis-)charge rate and the installed storage capacity.
+
+!!! warning "StorageBehavior"
+    `BoundRateTES` in its current implementation only supports
+    [`CyclicRepresentative`](@extref EnergyModelsBase.CyclicRepresentative) as storage behavior.
+    This input is not a required input due to the utilization of an inner constructor.
+
+# Fields
+- **`id`** is the name/identifier of the node.
+- **`level::AbstractStorageParameters`** are the level parameters of the `BoundRateTES`.
+  Depending on the chosen type, the level parameters can include variable OPEX and/or fixed OPEX.
+- **`stor_res::Resource`** is the stored [`Resource`](@extref EnergyModelsBase.Resource).
+- **`heat_loss_factor::Float64`** are the relative heat losses in percent.
+- **`level_discharge::Float64`** is the ratio of maximum discharge rate and installed storage level.
+- **`level_charge::Float64`** is the ratio of maximum charge rate and installed storage level.
+- **`input::Dict{<:Resource,<:Real}`** are the input [`Resource`](@extref EnergyModelsBase.Resource)s
+  with conversion value `Real`.
+- **`output::Dict{<:Resource,<:Real}`** are the generated [`Resource`](@extref EnergyModelsBase.Resource)s
+  with conversion value `Real`. Only relevant for linking and the stored
+  [`Resource`](@extref EnergyModelsBase.Resource) as the output value is not utilized in
+  the calculations.
+- **`data::Vector{<:Data}`** is the additional data (*e.g.*, for investments). The field `data`
+  is conditional through usage of a constructor.
+"""
+struct BoundRateTES{T} <: AbstractTES{T}
+    id::Any
+    level::EMB.UnionCapacity
+    stor_res::Resource
+    heat_loss_factor::Float64
+    level_charge::Float64
+    level_discharge::Float64
+    input::Dict{<:Resource,<:Real}
+    output::Dict{<:Resource,<:Real}
+    data::Vector{<:Data}
+end
+
+function BoundRateTES{T}(
+    id,
+    level::EMB.UnionCapacity,
+    stor_res::Resource,
+    heat_loss_factor::Float64,
+    level_charge::Float64,
+    level_discharge::Float64,
+    input::Dict{<:Resource,<:Real},
+    output::Dict{<:Resource,<:Real},
+) where {T<:EMB.StorageBehavior}
+    return BoundRateTES{T}(
+        id,
+        level,
+        stor_res,
+        heat_loss_factor,
+        level_charge,
+        level_discharge,
+        input,
+        output,
+        Data[],
+    )
+end
+
+function BoundRateTES(
+    id::Any,
+    level::EMB.UnionCapacity,
+    stor_res::Resource,
+    heat_loss_factor::Float64,
+    level_charge::Float64,
+    level_discharge::Float64,
+    input::Dict{<:Resource,<:Real},
+    output::Dict{<:Resource,<:Real},
+    data::Vector{<:Data},
+)
+    new{CyclicRepresentative}(
+        id,
+        level,
+        stor_res,
+        heat_loss_factor,
+        level_charge,
+        level_discharge,
         input,
         output,
         data,
@@ -318,7 +421,21 @@ end
 
 Returns the heat loss factor for storage `n`.
 """
-heat_loss_factor(n::ThermalEnergyStorage) = n.heat_loss_factor
+heat_loss_factor(n::AbstractTES) = n.heat_loss_factor
+
+"""
+    level_discharge(n::BoundRateTES)
+
+Returns the ratio of the maximum discharge rate and storage level capacity for TES `n`.
+"""
+level_discharge(n::BoundRateTES) = n.level_discharge
+
+"""
+    level_charge(n::BoundRateTES)
+
+Returns the ratio of the maximum charge rate and storage level capacity for TES `n`.
+"""
+level_charge(n::BoundRateTES) = n.level_charge
 
 """
     DirectHeatUpgrade