+++ NEWS.md
@@ -2,6 +2,9 @@
 
 ## Unversioned
 
+* Added Abstract type `AbstractTES`.
+* Introduced a new Storage type `BoundRateTES` including documentation, checks and tests which sets a maximum bound for charge and discharge capacities in relation to the installed storage level. 
+* Added discharge capacity to `ThermalEnergyStorage` node with legacy constructors to prevent breaking changes. 
 * Switched from `Data` to `ExtensionData` as described within [`EnergyModelsBase` v0.9.1](https://github.com/EnergyModelsX/EnergyModelsBase.jl/releases/tag/v0.9.1).
 * Minor fixes to the documentation and example comments.
 * Removed examples from test set as it somehow creates an error in the latest version on Ubuntu.